+++ package.json
@@ -19,8 +19,8 @@
   "dependencies": {
     "babel-runtime": "^5.8.20",
     "js-yaml": "^3.4.2",
+    "lodash": "^3.10.1",
     "swagger-parser": "^3.3.0",
-    "underscore": "^1.8.3",
     "yaml-js": "^0.1.3"
   },
   "devDependencies": {
