+++ test/fixtures/refract/payload-as-string.json
